[Refactor] DetectMobile 분리
몇몇 컴포넌트에서 복사해서 사용하던 DetectMobile을 Util로 분리
