+++ frontend/src/Components/Layout/SideMenuBtn/SideMenuBtnContainer.js
@@ -2,26 +2,12 @@ import React, { useLayoutEffect } from "react";
 import { useDispatch } from "react-redux";
 
 import { openedSideMenu } from "../../../Store/Actions/sideMenuAction";
+import detectMobile from "../../Util/DetectMobile";
 
 import SideMenuBtnPresenter from "./SideMenuBtnPresenter";
 
 const SideMenuBtnContainer = ({ showSideMenu, setShowSideMenu }) => {
   const dispatch = useDispatch();
-  const detectMobile = () => {
-    const toMatch = [
-      /Android/i,
-      /webOS/i,
-      /iPhone/i,
-      /iPod/i,
-      /BlackBerry/i,
-      /Windows Phone/i,
-    ];
-
-    return toMatch.some((toMatchItem) => {
-      return navigator.userAgent.match(toMatchItem);
-    });
-  };
-
   const handleSideMenuBtnClick = (event) => {
     event.preventDefault();
     setShowSideMenu(!showSideMenu);
